% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{inbreeding}
\alias{inbreeding}
\title{Inbreeding coefficients}
\usage{
inbreeding(pg)
}
\arguments{
\item{pg}{A graph from [pedigree_graph()].}
}
\value{
A double vector of `F`, one per input row, in input order.
}
\description{
Inbreeding coefficients
}
\examples{
# 5 is the child of full sibs 3 and 4.
pg <- pedigree_graph(data.frame(
  id = 1:5, mother = c(NA, NA, 1, 1, 3), father = c(NA, NA, 2, 2, 4)
))
inbreeding(pg)
}
